local node progress and overview. initial commit and cleanup
