+++ app/components/localNode/LocalNodeLog.js
@@ -1,13 +1,6 @@
 import React, { Component } from 'react';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import { smColors } from '/vars';
-import { BoldText, LeftPaneRow, ActionLink, BaseText } from './LocalNodeCommonComponents';
-
-type LogRecord = {
-  id: any,
-  text: string,
-  isReward: boolean
-};
 
 const getRandomNumber = (maxValue: number) => Math.floor(Math.random() * Math.floor(maxValue));
 const getTimestamp = () => new Date().toUTCString();
@@ -15,7 +8,43 @@ const generateLogEntry = (i: number) => {
   return { id: i, text: `${i} - ${getRandomNumber(i * 100)} - ${getTimestamp()} - Test log entry`, isReward: i % getRandomNumber(5) === 0 };
 };
 
-export const LogSection = styled.div`
+const Actionable = css`
+  &:hover {
+    opacity: 0.8;
+  }
+  &:active {
+    opacity: 0.6;
+  }
+`;
+
+const LeftPaneRow = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  padding: 22px 0;
+  border-bottom: 1px solid ${smColors.borderGray};
+  width: inherit;
+  height: 62px;
+`;
+
+const BaseText = styled.span`
+  font-size: 16px;
+  font-weight: normal;
+  line-height: 22px;
+`;
+
+const BoldText = styled(BaseText)`
+  font-weight: bold;
+`;
+
+const ActionLink = styled(BaseText)`
+  user-select: none;
+  color: ${smColors.green};
+  cursor: pointer;
+  ${Actionable}
+`;
+
+const LogSection = styled.div`
   height: 100%;
   display: flex;
   flex-direction: column;
@@ -24,14 +53,14 @@ export const LogSection = styled.div`
   position: relative;
 `;
 
-export const LogRowSection = styled.div`
+const LogRowSection = styled.div`
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   width: 100%;
 `;
 
-export const LogEntryWrapper = styled.div`
+const LogEntryWrapper = styled.div`
   height: 44px;
   line-height: 44px;
   flex: 1;
@@ -40,30 +69,36 @@ export const LogEntryWrapper = styled.div`
   text-overflow: ellipsis;
 `;
 
-export const LogRow = styled(LeftPaneRow)`
+const LogRow = styled(LeftPaneRow)`
   border-bottom: none;
   height: 100%;
   overflow-y: scroll;
   overflow-x: hidden;
 `;
 
-export const LogRowInner = styled.div`
+const LogRowInner = styled.div`
   height: 100%;
   width: 100%;
 `;
 
 // $FlowStyledIssue
-export const LogEntry = styled(BaseText)`
+const LogEntry = styled(BaseText)`
   color: ${({ reward }) => (reward ? smColors.green : smColors.black)};
 `;
 
-export const BottomGradient = styled.div`
+const BottomGradient = styled.div`
   position: absolute;
   width: 100%;
   height: 44px;
   bottom: 0;
   background-image: linear-gradient(transparent, white);
 `;
+
+type LogRecord = {
+  id: any,
+  text: string,
+  isReward: boolean
+};
 
 class LocalNodeLog extends Component<{}, { log: LogRecord[] }> {
   timer: any;
