+++ app/components/localNode/index.js
@@ -1,24 +1,6 @@
-export {
-  BaseText,
-  RightHeaderText,
-  BaseImage,
-  RightPaneInner,
-  LeftPaneInner,
-  ActionLink,
-  ImageWrapper,
-  BottomPaddedRow,
-  ItemTextWrapperUniformPadding,
-  BottomLinksWrapper,
-  CenterTextWrapper,
-  LinkTextWrapper,
-  ItemText,
-  BoldText,
-  GrayText
-} from './LocalNodeCommonComponents';
 export { default as LoadingBar } from './LoadingBar';
 export { default as LocalNodeLog } from './LocalNodeLog';
 export { default as LocalNodeStatus } from './LocalNodeStatus';
-export { default as LocalNodeBase } from './LocalNodeBase';
-export { default as RightPaneSetup } from './RightPaneSetup';
 export { default as RightPane } from './RightPane';
 export { default as LeftPaneSetup } from './LeftPaneSetup';
+export { default as LeftPane } from './LeftPane';
